requirements.txt
# Core dependencies for Conversation Analytics Platform
duckdb>=1.0.0
boto3>=1.35.50
urllib3>=2.0.0
pytz>=2024.1
python-dotenv>=1.0.0
rich>=13.0.0
pandas>=2.0.0

# AI/ML dependencies
google-generativeai>=0.8.0
openai>=1.0.0

# Additional dependencies that may be required
httpx>=0.25.0
requests>=2.31.0